NovelPropsの定義場所修正とimportパス修正
