+++ src/components/novelCard/NovelCard.tsx
@@ -5,39 +5,6 @@ import CommentButton from '../buttonicon/CommentButton';
 import NextPlanButton from '../buttonicon/NextPlanButton';
 import { Sentence } from '../../types/sentences';
 
-export interface NovelProps {
-  title: string;
-  main_copy: string;
-  overview: string;
-  popular: boolean;
-  newArrival: boolean;
-  author_user_name: string;
-  chips: { label: string }[];
-  tags: { label: string }[];
-  reader_count: number;
-  avatar: {
-    src: string;
-    alt: string;
-    color: string;
-    text: string;
-  };
-  sentence_id: number;
-  sentence_user_count: number;
-  sentence_hierarchy_count: number;
-  sentence: string;
-  textIndex: number;
-  userId: number;
-  userName: string;
-  profile_icon_image: string;
-  evaluation_good_count: number;
-  evaluation_stay_count: number;
-  created_at: string;
-  updated_at: string;
-  children: Sentence[];
-  main: Sentence[];
-  parent: Sentence[];
-}
-
 interface NovelCardProps {
   key: number;
   index: number;
